+++ examples/directions.rs
@@ -28,39 +28,39 @@ fn main() {
     let mut r = Raster::new(p.width(), p.height());
 
     // Render paths.
-    let path = font.render(english2, (64.0, 0.0, 512.0 - 64.0, 512.0 - FONT_SIZE), (FONT_SIZE, FONT_SIZE), font::TextAlign::Left);
+    let path = font.render(english2, (64.0, 0.0, 512.0 - 64.0, 512.0 - FONT_SIZE), (FONT_SIZE, FONT_SIZE), font::TextAlign::Left).0;
     let path: Vec<font::PathOp> = path.collect();
     r.over(p.fill(&path, FillRule::NonZero), Rgba8::rgb(0, 0, 0));
 
-    let path = font.render(nepali, (64.0, 32.0 * 1.0, 512.0 - 64.0, 512.0 - FONT_SIZE * 2.0), (FONT_SIZE, FONT_SIZE), font::TextAlign::Left);
+    let path = font.render(nepali, (64.0, 32.0 * 1.0, 512.0 - 64.0, 512.0 - FONT_SIZE * 2.0), (FONT_SIZE, FONT_SIZE), font::TextAlign::Left).0;
     let path: Vec<font::PathOp> = path.collect();
     r.over(p.fill(&path, FillRule::NonZero), Rgba8::rgb(0, 0, 0));
 
-    let path = font.render(english, (64.0, 32.0 * 2.0, 512.0 - 64.0, 512.0 - 32.0 * 5.0), (FONT_SIZE, FONT_SIZE), font::TextAlign::Center);
+    let path = font.render(english, (64.0, 32.0 * 2.0, 512.0 - 64.0, 512.0 - 32.0 * 5.0), (FONT_SIZE, FONT_SIZE), font::TextAlign::Center).0;
     let path: Vec<font::PathOp> = path.collect();
     r.over(p.fill(&path, FillRule::NonZero), Rgba8::rgb(0, 0, 0));
 
-    let path = font.render(arabic, (64.0, 32.0 * 3.0, 512.0 - 64.0, 512.0 - FONT_SIZE * 3.0), (FONT_SIZE, FONT_SIZE), font::TextAlign::Right);
+    let path = font.render(arabic, (64.0, 32.0 * 3.0, 512.0 - 64.0, 512.0 - FONT_SIZE * 3.0), (FONT_SIZE, FONT_SIZE), font::TextAlign::Right).0;
     let path: Vec<font::PathOp> = path.collect();
     r.over(p.fill(&path, FillRule::NonZero), Rgba8::rgb(0, 0, 0));
 
-    let path = font.render(hebrew, (64.0, 32.0 * 4.0, 512.0 - 64.0, 512.0 - 32.0 * 4.0), (FONT_SIZE, FONT_SIZE), font::TextAlign::Right);
+    let path = font.render(hebrew, (64.0, 32.0 * 4.0, 512.0 - 64.0, 512.0 - 32.0 * 4.0), (FONT_SIZE, FONT_SIZE), font::TextAlign::Right).0;
     let path: Vec<font::PathOp> = path.collect();
     r.over(p.fill(&path, FillRule::NonZero), Rgba8::rgb(0, 0, 0));
 
-    let path = font.render(nepali, (64.0, 32.0 * 5.0, 512.0 - 64.0, 512.0 - 32.0 * 6.0), (FONT_SIZE, FONT_SIZE), font::TextAlign::Center);
+    let path = font.render(nepali, (64.0, 32.0 * 5.0, 512.0 - 64.0, 512.0 - 32.0 * 6.0), (FONT_SIZE, FONT_SIZE), font::TextAlign::Center).0;
     let path: Vec<font::PathOp> = path.collect();
     r.over(p.fill(&path, FillRule::NonZero), Rgba8::rgb(0, 0, 0));
 
-    let path = font.render(english, (64.0, 32.0 * 6.0, 512.0 - 64.0, 512.0 - FONT_SIZE), (FONT_SIZE, FONT_SIZE * 2.0), font::TextAlign::Right);
+    let path = font.render(english, (64.0, 32.0 * 6.0, 512.0 - 64.0, 512.0 - FONT_SIZE), (FONT_SIZE, FONT_SIZE * 2.0), font::TextAlign::Right).0;
     let path: Vec<font::PathOp> = path.collect();
     r.over(p.fill(&path, FillRule::NonZero), Rgba8::rgb(0, 0, 0));
 
-    let path = font.render(korean, (0.0, 0.0, 512.0, 512.0 - 32.0 * 7.0), (FONT_SIZE, FONT_SIZE), font::TextAlign::Vertical);
+    let path = font.render(korean, (0.0, 0.0, 512.0, 512.0 - 32.0 * 7.0), (FONT_SIZE, FONT_SIZE), font::TextAlign::Vertical).0;
     let path: Vec<font::PathOp> = path.collect();
     r.over(p.fill(&path, FillRule::NonZero), Rgba8::rgb(0, 0, 0));
 
-    let path = font.render(japanese, (32.0, 0.0, 512.0, 512.0 - 32.0 * 7.0), (FONT_SIZE, FONT_SIZE), font::TextAlign::Vertical);
+    let path = font.render(japanese, (32.0, 0.0, 512.0, 512.0 - 32.0 * 7.0), (FONT_SIZE, FONT_SIZE), font::TextAlign::Vertical).0;
     let path: Vec<font::PathOp> = path.collect();
     r.over(p.fill(&path, FillRule::NonZero), Rgba8::rgb(0, 0, 0));